billie_pr/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filter==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.11.1
importlib-metadata==2.0.0
Markdown==3.2.2
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
zipp==3.2.0
